add roles and decorators

+++ src/restaurants/restaurants.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, Req, UploadedFiles, UseGuards, UseInterceptors } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, Param, Post, Put, Query, Req, UploadedFiles, UseGuards, UseInterceptors } from '@nestjs/common';
 import { FilesInterceptor } from '@nestjs/platform-express';
 import { Query as ExpressQuery } from 'express-serve-static-core';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
@@ -8,6 +8,8 @@ import { Restaurant } from './schemas/restaurant.schema';
 import { AuthGuard } from '@nestjs/passport';
 import { CurrentUser } from 'src/auth/decorators/current-user.decorator';
 import { User } from 'src/auth/schemas/user.schema';
+import { RolesGuard } from 'src/auth/guards/roles.guards';
+import { Roles } from 'src/auth/decorators/roles.decorator';
 
 @Controller('restaurants')
 export class RestaurantsController {
@@ -28,21 +30,26 @@ export class RestaurantsController {
     }
 
     @Post()
-    @UseGuards(AuthGuard())
+    @UseGuards(AuthGuard(), RolesGuard)
     async createRestaurant(@Body() restaurant: CreateRestaurantDto, @CurrentUser() user: User
     ): Promise<Restaurant> {
         return this.restaurantsServices.create(restaurant, user)
     }
 
     @Put(':id')
-    async updateRestaurant(@Param('id') id: string, @Body() restaurant: UpdateRestaurantDto): Promise<Restaurant> {
-        await this.restaurantsServices.findById(id)
-        return this.restaurantsServices.updateById(id, restaurant)
+    @UseGuards(AuthGuard(), RolesGuard)
+    async updateRestaurant(@Param('id') id: string, @Body() restaurantDto: UpdateRestaurantDto, @CurrentUser() user: User): Promise<Restaurant> {
+        const restaurant = await this.restaurantsServices.findById(id)
+        console.log(user)
+        if (restaurant.user.toString() !== user._id.toString()) throw new ForbiddenException('You can not update this restaurant')
+        return this.restaurantsServices.updateById(id, restaurantDto)
     }
 
     @Delete(':id')
-    async deleteRestaurant(@Param('id') id: string): Promise<{ deleted: Boolean }> {
+    @UseGuards(AuthGuard(), RolesGuard)
+    async deleteRestaurant(@Param('id') id: string, @CurrentUser() user: User): Promise<{ deleted: Boolean }> {
         const restaurant = await this.restaurantsServices.findById(id)
+        if (restaurant.user.toString() !== user._id.toString()) throw new ForbiddenException('You can not delete this restaurant')
         const isDeleted = await this.restaurantsServices.deleteImages(restaurant.images)
         if (isDeleted) {
             await this.restaurantsServices.deleteById(id)
